+++ src/settings/JobProfiles/JobProfileDetailsActionMenu/index.js
@@ -1,0 +1,1 @@
+export { default as JobProfileDetailsActionMenu } from './JobProfileDetailsActionMenu';